api/requirements.txt
pytest
pytest-emoji
pytest-html
pytest-md
pytest-parallel
pytest-repeat
python-multipart
fastapi
uvicorn
mangum
bs4
lxml
nltk
requests
pandas
scikit-learn
azure-storage-blob
jinja2
ruamel.yaml
